Getting enet working requires version 10.0.0.07
When using the MCU+ SDK version 9.1.0.05 sysconfig will, for some
reason, ignore any value given to CPSW_CPTS_RFTCLK_FREQ and generate
output as if it was set to 500MHz. This will cause an assertion on
bootup to fail as it tries to assert(250MHz % 500MHz == 0), which will
always fail.